terminar de poner los float 2

+++ RAPPI2/src/Interaccion/Calificacion.java
@@ -10,7 +10,7 @@ package Interaccion;
 
 public class Calificacion {
 	private Object calificado;
-	private double puntuacion= 5.0;
+	private float puntuacion= 5.0;
 	private Object calificador;
 	
 	/**
@@ -19,16 +19,16 @@ public class Calificacion {
 	 * @param puntuacion El parametro puntuacion define la puntuación que le dio el calificador a el calificado
 	 * @param calificador El parametro calificador define el objeto que realizo la calificacion
 	 */
-	public Calificacion(Object calificado,double puntuacion,Object calificador) {
+	public Calificacion(Object calificado,float puntuacion,Object calificador) {
 		this.calificado = calificado;
 		this.calificador = calificador;
 		this.puntuacion = puntuacion;
 	}
 	
-	public double getPuntuacion() {
+	public float getPuntuacion() {
 		return this.puntuacion;
 	}
-	public void setPuntuacion(int puntuacion) {
+	public void setPuntuacion(float puntuacion) {
 		
 		this.puntuacion = puntuacion;
 	}
